Changed Python SCL setup.cfg name to "stepcode-scl" so the package can be deployed to pypi.
The current package name, SCL, is too similar to existing pypi package names that the repository will not allow it to be committed. This change does not affect internal package name, which remains "SCL", just the name used to fetch the package from the pypi repository.

After this change is committed and posted to pypi, users will be able to fetch the package with `pip install stepcode-scl`.